refactor: refactor api and app code into respective folders
