+++ electron-src/api/index.ts
@@ -1,8 +1,4 @@
-/**
- * anime - info[malId] / search[query] / posters / recommended / genre[genre]
- * episode - source[animeId, episodeNum] / info[animeId, epStart, epEnd?]
- */
-import * as anime from "./anime"
-import * as episode from "./episode"
-
-export const api = { anime, episode };
+// register all the ipc functions
+import "./anime/ipc";
+import "./episode/ipc";
+import "./other/ipc";
