+++ electron-src/types.ts
@@ -1,17 +1,7 @@
-import { Anime } from "@prisma/client";
+import { Anime, Episode } from "@prisma/client";
 
 export type WindowState = "maximise" | "minimize" | "close";
 
 export interface EpisodeWithAnime extends Episode {
   anime: Anime;
-}
-
-export interface Episode {
-  watchTime: number;
-  number: number;
-  title: string;
-  animeId: string;
-  length: number;
-  source: string;
-  lastUpdated: Date;
 }
