+++ package.json
@@ -19,7 +19,7 @@
     "test": "jest",
     "svelte:build": "cd svelte-src && yarn build",
     "dev": "concurrently -n=svelte,electron -c='#ff3e00',blue \"npm run svelte:dev\" \"npm run electron:dev\"",
-    "electron:start": "tsc && electron dist | pino-pretty",
+    "electron:start": "tsc && electron dist/app | pino-pretty",
     "electron:dev": "nodemon",
     "build": "npm run prepublish && tsc && electron-builder",
     "prepare": "npx prisma generate && npm run svelte:install && npm run svelte:build && npx prisma generate && tsc",
@@ -85,7 +85,6 @@
     "electron-updater": "^5.3.0",
     "fast-xml-parser": "^4.0.11",
     "fluent-ffmpeg": "^2.1.2",
-    "m3u8stream": "^0.8.6",
     "prisma": "^4.6.1",
     "string-similarity": "^4.0.4",
     "winston": "^3.8.2"
