+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["electron-src/index.ts", "electron-src/preload.ts"],
+  "include": ["electron-src/app/index.ts", "electron-src/app/preload.ts"],
   "exclude": ["svelte-src/src/*/**"],
   "compilerOptions": {
     "outDir": "dist",
